master: no message

+++ lib-build.sh
@@ -193,25 +193,21 @@ function mavenBuild()
     #__mvn_jar_filter=$(mvn help:evaluate -Dexpression=project.build.finalName -q -DforceStdout)
     __mvn_jar_filter="app-0.0.1-SNAPSHOT.jar"
     #binary jar file name
-    __mvn_jar_source_file=${__mvn_build_src_bin_dir}/${__mvn_jar_filter}.jar
-    echo "__mvn_jar_source_file==${__mvn_jar_source_file}"
-    ls -l ${__mvn_jar_source_file}
-    echo "__mvn_jar_source_file==${__mvn_jar_source_file}"
-    echo "__mvn_jar_filter==${__mvn_jar_filter}"
+    __mvn_jar_source_file=${__mvn_build_src_bin_dir}/${__mvn_jar_filter}
     echG "      - jar file: ${__mvn_jar_source_file}"
     read
-    # if ! [[ -f ${__mvn_jar_source_file} ]]; then
-    #   echY "      jar file: ${__mvn_jar_source_file}"
-    #   echR "      ==============================  "
-    #   echR "      ******JAR file not found******  "
-    #   echR "      ******JAR file not found******  "
-    #   echR "      ==============================  "
-    # else
+    if ! [[ -f ${__mvn_jar_source_file} ]]; then
+      echY "      jar file: ${__mvn_jar_source_file}"
+      echR "      ==============================  "
+      echR "      ******JAR file not found******  "
+      echR "      ******JAR file not found******  "
+      echR "      ==============================  "
+    else
       __mvn_jar_source_file_new=${__mvn_build_base_dir}/$(basename ${__mvn_jar_source_file})
       mv ${__mvn_jar_source_file} ${__mvn_jar_source_file_new}
       export __func_return="${__func_return} ${__mvn_jar_source_file_new}"
       echC "      - JAR file: ${__mvn_jar_source_file_new}"
-    #fi
+    fi
   fi
   cd ${__mvn_build_base_dir}
   rm -rf ${__mvn_build_src_dir}
